Average Salary
You are given an array of unique integers salary where salary[i] is the salary of the ith employee.

Return the average salary of employees excluding the minimum and maximum salary. Answers within 10-5 of the actual answer will be accepted.

In this code, we first check if the input array has at least three elements (minimum, maximum, and one more to calculate the average). If there are fewer than three elements, we return 0 as the average cannot be calculated.

Next, we sort the array of salaries in ascending order using std::sort. Then, we iterate from the second element (salary[1]) to the second-to-last element (salary[n - 2]), summing up all the salaries excluding the minimum and maximum.

Finally, we calculate the average by dividing the sum by the number of elements minus 2 (n - 2). We use static_cast<double>(sum) to perform a floating-point division instead of integer division.